requirements.txt
asgiref==3.6.0
Django==4.2.6
django-debug-toolbar==4.2.0
django-multiselectfield==0.1.12
Pillow==10.0.0
psycopg2-binary==2.9.7
pymemcache==4.0.0
python-memcached==1.59
six==1.16.0
sqlparse==0.4.3
